ApplicationLibrary/mqtt-mnist/mqtt-cnn-mnist-infer/requirements.txt
torch
torchvision
torchaudio
Pillow
paho-mqtt